# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ctrlX System Identification API",
...
    "contact": {
      "name": "Equipo ctrlX SysID"
    },
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/identification/run": {
      "post": {
        "tags": [
          "Identificación"
        ],
        "summary": "Re-identificar con el buffer actual",
        "description": "Vuelve a correr la identificación sobre las muestras que ya están en el\nbuffer, usando las **condiciones de ensayo vigentes**\n(`GET /api/test/config`).\n\nLa identificación automática solo se dispara una vez por escalón, y espera\na tener la ventana completa que pide `duration_s`. Este endpoint es el\ndisparador manual: sirve para recalcular tras cambiar el `order` o la\nescala de una señal, sin repetir el ensayo en el PLC.\n\nCon `allow_partial=true` (por defecto) identifica con las muestras que\nhaya, aunque sean menos que las que pide la duración configurada, y marca\nel resultado con `truncated: true`. Un proceso de primer orden queda bien\ndefinido a partir de unas 3-4 constantes de tiempo, así que una ventana\nmás corta que el ensayo suele bastar; el flag está para que se sepa que el\najuste se hizo sobre menos datos de los previstos.\n\nCon `allow_partial=false` exige la ventana completa y devuelve 400 si\ntodavía no está.\n\nDevuelve 400 si no hay escalón en el buffer o si ningún modelo converge.",
        "operationId": "run_identification_api_identification_run_post",
        "parameters": [
          {
            "name": "allow_partial",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Identificar aunque la ventana sea más corta que la duración configurada.",
              "default": true,
              "title": "Allow Partial"
            },
            "description": "Identificar aunque la ventana sea más corta que la duración configurada."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentificationResult"
                }
              }
            }
          },
          "400": {
            "description": "El buffer actual no permite identificar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__schemas__opcua__ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/test/scales": {
      "get": {
        "tags": [
          "Ensayo"
        ],
        "summary": "Escala vigente por rol",
        "description": "`ma` = 4-20 mA, `pct` = 0-100 %, `v` = 0-10 V.\n\nInternamente el backend trabaja en **% de span**: es la única escala en la que la ganancia del modelo es adimensional y las sintonías PID son comparables entre lazos. La escala declarada aquí es la que se usa para traducir cada lectura cruda del PLC a ese %.",
        "operationId": "get_scales_api_test_scales_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScalesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ensayo"
        ],
        "summary": "Fijar la escala de cada rol",
        "description": "Cambia en qué escala física se interpreta la variable de cada rol. Los roles omitidos conservan su valor.\n\nLos campos `actuator_pct`, `sensor_pct` y `setpoint_pct` se recalculan **también para las muestras que ya están en el buffer**, de modo que un ensayo ya capturado queda reinterpretado con la escala nueva y no hace falta repetirlo.\n\nNo se escribe nada en el PLC.",
        "operationId": "set_scales_api_test_scales_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScalesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScalesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__schemas__simulation__ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/test/config": {
      "get": {
        "tags": [
          "Ensayo"
        ],
        "summary": "Condiciones del ensayo",
        "description": "Valor inicial y final, retardo, duración, orden y todos los derivados.",
        "operationId": "get_config_api_test_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestConfigResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ensayo"
        ],
        "summary": "Guardar las condiciones del ensayo",
        "description": "Guarda el escalón que el operador va a aplicar. Los campos omitidos conservan su valor anterior.\n\nEl backend usa estos números para tres cosas:\n\n1. **Umbral de detección**: se considera escalón un salto de al menos la mitad del `delta_pct` configurado, con un piso de 1 %. Así no se dispara con ruido ni se pierde el evento si el actuador no llega exacto al valor final.\n2. **Ventana de identificación**: `delay_s` define cuántas muestras de línea base se conservan antes del salto y `duration_s - delay_s` cuántas de respuesta se capturan después.\n3. **Modelo**: si `order` no es `auto`, se ajusta solo ese modelo.\n\nDevuelve 400 si el escalón no es identificable (valores fuera de escala, salto nulo, o muy pocas muestras después del escalón).",
        "operationId": "set_config_api_test_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestConfigResponse"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__schemas__simulation__ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/test/convert": {
      "post": {
        "tags": [
          "Ensayo"
        ],
        "summary": "Convertir un valor entre escalas",
        "description": "Conversión lineal por span, pasando siempre por porcentaje:\n\n```\n%     = (valor - min) / (max - min) * 100\nvalor = min + % / 100 * (max - min)\n```\n\nSirve para los textos de ayuda del paso 2 (`8.0 mA = 25 %`) sin duplicar la fórmula en el JavaScript.",
        "operationId": "convert_api_test_convert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConvertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConvertResponse"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__schemas__simulation__ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/test/plan": {
      "get": {
        "tags": [
          "Ensayo"
        ],
        "summary": "Perfil del actuador, muestra a muestra",
        "description": "Devuelve el perfil que el backend va a comandar: **una entrada por muestra** del ensayo, al mismo periodo con el que se ejecutará.\n\nNo confundir con `/preview`, que reparte N puntos solo para dibujar la vista previa del paso 2. Aquí `time[i]` es el instante real de la muestra `i`.\n\nSi hay un ensayo en curso devuelve el plan de ESE ensayo (congelado al arrancar); si no, el que se generaría con la configuración actual.",
        "operationId": "get_plan_api_test_plan_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestPlan"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__schemas__simulation__ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/test/start": {
      "post": {
        "tags": [
          "Ensayo"
        ],
        "summary": "Arrancar el ensayo",
        "description": "Arranca el reloj del ensayo en el backend y empieza a emitir eventos por WebSocket: `test_started` con el plan completo, luego un `test_tick` por muestra, y `test_finished` al terminar.\n\n**El reloj vive en el backend, no en el navegador.** Así el ensayo sobrevive a que se cierre la pestaña y no depende de los timers del navegador, que se estrangulan en segundo plano.\n\nPor defecto limpia el buffer de muestras y descarta la identificación anterior, para que el ensayo nuevo no arrastre datos del anterior.\n\n**En esta fase no se escribe nada en el PLC**: el runner solo calcula y publica el valor comandado. El operador sigue aplicando el escalón desde el ctrlX. `writes_enabled` indica si eso ya cambió.\n\nDevuelve 400 si ya hay un ensayo corriendo o si la configuración del paso 2 no es válida.",
        "operationId": "start_test_api_test_start_post",
        "parameters": [
          {
            "name": "clear_buffer",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Vaciar el buffer y la identificación previa antes de arrancar.",
              "default": true,
              "title": "Clear Buffer"
            },
            "description": "Vaciar el buffer y la identificación previa antes de arrancar."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestStartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__schemas__simulation__ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/test/writer": {
      "get": {
        "tags": [
          "Ensayo"
        ],
        "summary": "Estado de la escritura al PLC",
        "description": "Dice si el ensayo va a **mover el actuador** o solo dibujar, y si la variable mapeada admite escritura.",
        "operationId": "get_writer_api_test_writer_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriterResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ensayo"
        ],
        "summary": "Armar o desarmar la escritura al PLC",
        "description": "**Mientras no se arme, el ensayo no toca el PLC**: calcula el perfil y lo publica, pero el actuador no se mueve. Armar es un paso aparte a propósito, para que pulsar *Inicio* no mueva equipo por accidente.\n\nAl armar se comprueba que haya sesión OPC UA, que el rol `actuator` tenga una variable asignada y que el servidor la declare escribible. Si algo falla se devuelve 400 y la escritura queda desarmada.\n\nSe escribe sobre **la misma variable que se eligió en el paso 1**: la que esté mapeada al rol `actuator` en ese momento.\n\nNo se puede armar ni desarmar con un ensayo en curso: cambiar quién gobierna el actuador a mitad de camino dejaría un registro donde una parte del escalón se aplicó y otra no.",
        "operationId": "set_writer_api_test_writer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__schemas__simulation__ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/test/preview": {
      "get": {
        "tags": [
          "Ensayo"
        ],
        "summary": "Vista previa del escalón",
        "description": "Serie ideal del escalón configurado, lista para graficar en el paso 2.\n\nCon `scale` se puede pedir en otra escala (por ejemplo en `pct` aunque el actuador esté en `ma`) sin cambiar la configuración guardada.",
        "operationId": "preview_api_test_preview_get",
        "parameters": [
          {
            "name": "points",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 2,
              "description": "Cantidad de puntos a generar",
              "default": 200,
              "title": "Points"
            },
            "description": "Cantidad de puntos a generar"
          },
          {
            "name": "scale",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Escala de salida. Por defecto, la del actuador.",
              "title": "Scale"
            },
            "description": "Escala de salida. Por defecto, la del actuador."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IdentificationResult": {
        "properties": {
          "step_index": {
            "type": "integer",
            "title": "Step Index",
            "description": "Índice del escalón dentro de la serie completa"
          },
          "order": {
            "type": "string",
            "title": "Order",
            "description": "Orden pedido en las condiciones de ensayo. Con `auto` vienen los tres modelos rankeados; con `fopdt`/`sopdt`/`integrating`, `models` trae uno solo.",
            "default": "auto"
          },
          "winner": {
            "type": "string",
            "title": "Winner",
            "description": "model_type del modelo con mejor R²"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "description": "True si la ventana usada fue más corta que la que pide `duration_s`. El ajuste sigue siendo válido, pero se hizo sobre menos respuesta de la prevista: conviene revisar que la curva haya llegado al nuevo estable.",
            "default": false
          },
          "requested_post_samples": {
            "type": "integer",
            "title": "Requested Post Samples",
            "description": "Muestras de respuesta que pedía la duración configurada",
            "default": 0
          },
          "window": {
            "$ref": "#/components/schemas/IdentificationWindow"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelResult"
            },
            "type": "array",
            "title": "Models",
            "description": "Ordenados de mejor a peor R²"
          }
        },
        "type": "object",
        "required": [
          "step_index",
          "winner",
          "window",
          "models"
        ],
        "title": "IdentificationResult",
        "example": {
          "models": [],
          "order": "auto",
          "step_index": 30,
          "window": {
            "actuator": [
              4.0,
              4.0,
              12.0
            ],
            "count": 3,
            "sensor": [
              8.0,
              8.0,
              8.31
            ],
            "setpoint": [
              12.0,
              12.0,
              12.0
            ],
            "time": [
              0.0,
              0.5,
              1.0
            ]
          },
          "winner": "fopdt"
        }
      },
...
    }
  },
...
}